requirements.txt
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.33
langchain_groq==0.1.9
langchain_nomic==0.1.2
pinecone==5.0.1
pinecone_text==0.9.0
python-dotenv==1.0.1
streamlit==1.37.1
